bug in deck calibration. root cause is that the  field in api was a str but the fastapi passed in an enum. the  str is a key in a dict and the lookup using the enum failed due to a str conversion before lookup. solution is to move the enum into the API and have it referenced by fastapi model. Also, remove the str conversion before look up.
